Helping stuff for debugging https://svn.eionet.europa.eu/projects/Reportnet/ticket/1899

+++ src/main/java/eionet/cr/harvest/PullHarvest.java
@@ -8,6 +8,7 @@ import java.net.URL;
 import java.net.URLConnection;
 import java.util.Date;
 
+import eionet.cr.config.GeneralConfig;
 import eionet.cr.util.FileUtil;
 
 /**
@@ -36,44 +37,59 @@ public class PullHarvest extends Harvest{
 	 */
 	protected void doExecute() throws HarvestException{
 
-		File toFile = fullFilePathForSourceUrl(sourceUrlString);;
-		if (toFile.exists())
-			toFile.delete();
+		boolean doDownload = true;
 		
-		logger.debug("Downloading");
+		File toFile = fullFilePathForSourceUrl(sourceUrlString);;
+		if (toFile.exists()){
+			doDownload = !Boolean.parseBoolean(GeneralConfig.getProperty(GeneralConfig.HARVESTER_DEBUG_USE_DOWNLOADED_FILES, "true"));
+			if (doDownload){
+				toFile.delete();
+			}
+		}
 		
 		InputStream inputStream = null;
 		try{
-			sourceAvailable = new Boolean(false);
 			
-			URL url = new URL(sourceUrlString);
-			URLConnection urlConnection = url.openConnection();
-			
-			urlConnection.setRequestProperty("Accept", "application/rdf+xml, text/xml, */*");
-			if (lastHarvest!=null){
-				urlConnection.setIfModifiedSince(lastHarvest.getTime());
-			}
-			
-			inputStream = urlConnection.getInputStream();
-			sourceAvailable = new Boolean(true);
-			
-			if (urlConnection instanceof HttpURLConnection){
-				if (((HttpURLConnection)urlConnection).getResponseCode()==HttpURLConnection.HTTP_NOT_MODIFIED){
-					String msg = "Source not modified since " + lastHarvest.toString();
-					logger.debug(msg);
-					infos.add(msg);
-					return;
+			if (doDownload){
+				
+				logger.debug("Downloading");
+				
+				sourceAvailable = Boolean.FALSE;
+				
+				URL url = new URL(sourceUrlString);
+				URLConnection urlConnection = url.openConnection();
+				
+				urlConnection.setRequestProperty("Accept", "application/rdf+xml, text/xml, */*");
+				if (lastHarvest!=null){
+					urlConnection.setIfModifiedSince(lastHarvest.getTime());
 				}
-			}
+				
+				inputStream = urlConnection.getInputStream();
+				sourceAvailable = Boolean.TRUE;
+				
+				if (urlConnection instanceof HttpURLConnection){
+					if (((HttpURLConnection)urlConnection).getResponseCode()==HttpURLConnection.HTTP_NOT_MODIFIED){
+						String msg = "Source not modified since " + lastHarvest.toString();
+						logger.debug(msg);
+						infos.add(msg);
+						return;
+					}
+				}
 			
-			FileUtil.streamToFile(inputStream, toFile);
+				// harvest the downloaded file
+				FileUtil.streamToFile(inputStream, toFile);
+			}
+			else{
+				sourceAvailable = Boolean.TRUE;
+			}
 		}
 		catch (IOException e){
 			throw new HarvestException(e.toString(), e);
 		}
 		finally{
 			try{
-				if (inputStream!=null) inputStream.close();
+				if (inputStream!=null)
+					inputStream.close();
 			}
 			catch (IOException e){}
 		}
